requirements.txt
bottle
pillow
cryptography
Werkzeug
google-api-python-client
google-auth-httplib2
google-auth-oauthlib